server-java-modern: fix all 24 lossy-conversion warnings
Three categories of implicit lossy primitive conversions resolved:

* World.java traversalMask (15 sites): TileValue.traversalMask is
  declared `byte` (CollisionFlag uses bits 0-7 only, fits in 8 bits)
  but the boundary/scenery clear-bit code uses 16-bit-shaped masks
  like 0xffbf, 0xfffd, etc. The masks work correctly via int promote
  + truncate, but the implicit (int -> byte) cast was lint-flagged.
  Made the truncation explicit with `(byte)` so it's clear at the
  call site that we're discarding the high 8 bits intentionally.

* BCrypt (3 sites): two `byte o |= int` and one `byte ret |= int^int`
  in encode_base64 and equalsNoEarlyReturn. Wrapped the rhs in
  `(byte)`; same explicit-truncation pattern.

* Equipment.armourBonus (2 sites): `total += long` where total is
  int, sourced from ItemDefinition.getArmourBonus() which returns
  long. Bonuses fit in int in practice (typical values 0..100), so
  truncation is safe; made `(int)` explicit at the call sites.

* Combat / skill multipliers (4 sites): `damage += maximum*0.2`,
  `fishCaught *= 1.5`, `timeToCook *= 0.7` — all `int *= double`
  forms where the implicit truncation hid behind compound assignment.
  Rewrote as `x = (int) (x * factor)` so the float-then-truncate is
  explicit.

Verified: ant compile clean, smoke_test.sh 4/4 pass, no remaining
lossy-conversion warnings.

+++ server-java-modern/src/com/openrsc/server/event/rsc/impl/combat/OSRSCombatFormula.java
@@ -164,7 +164,7 @@ public class OSRSCombatFormula {
 				// Check if strength cape should activate
 				final int maxHit = calcMaxHit(attacker);
 				if (damage >= maxHit - (maxHit * 0.5) && SkillCapes.shouldActivate((Player) attacker, STRENGTH_CAPE, isHit)) {
-					damage += (maxHit*0.2);
+					damage += (int) (maxHit * 0.2);
 					((Player) attacker).message("@ora@Your Strength cape has granted you a critical hit");
 				}
 			}
